is work bot clean arch

+++ internal/domain/models/user.go
@@ -1,12 +1,12 @@
 package models
 
 type User struct {
-	ID     		int64
+	ID     		int
 	ChatID 		string
 	MessageTime string
 }
 
-func New(id int64, chatID, messTime string) *User {
+func New(id int, chatID, messTime string) *User {
 	return &User{
 		ID:          id,
 		ChatID: 	 chatID,
